Remove psservice package and merge it into pstore server
