+++ cmd/farmer/main.go
@@ -11,7 +11,7 @@ import (
 
 	"storj.io/storj/pkg/cfgstruct"
 	"storj.io/storj/pkg/kademlia"
-	"storj.io/storj/pkg/piecestore/psservice"
+	psserver "storj.io/storj/pkg/piecestore/rpc/server"
 	"storj.io/storj/pkg/process"
 	"storj.io/storj/pkg/provider"
 )
@@ -35,7 +35,7 @@ var (
 	runCfg struct {
 		Identity provider.IdentityConfig
 		Kademlia kademlia.Config
-		Storage  psservice.Config
+		Storage  psserver.Config
 	}
 	setupCfg struct {
 		BasePath string `default:"$CONFDIR" help:"base path for setup"`
